reorganized logging and improved aliases
